requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
langdetect>=1.0.9
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.11
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
wikipedia-api>=0.6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0.1
scikit-learn
wikipedia 